Service, cache and UI refactor (#2)
* Update DTOs
* Add impact radius
* Set agm-circle
* Change cache strategy
* Update image cache
* Sync on refresh
* Add drawer modal
* Smooth zoom

+++ src/app/auth/auth.component.ts
@@ -7,11 +7,7 @@ import {ToastrService} from 'ngx-toastr';
 import {JwtHelperService} from '@auth0/angular-jwt';
 import {SessionService} from '../service/session.service';
 import {CustomReuseStrategy} from '../main/common/custom.reuse.strategy';
-import {SpinnerService} from '../shared/spinner/spinner.service';
-import {concatMap} from 'rxjs/operators';
-import {LoginRequest} from '../model/request/login.request';
-import {RegisterRequest} from '../model/request/register.request';
-import {JWT_OFFSET_SECONDS, MAX_EMAIL_LENGTH, MAX_PASSWORD_LENGTH, MIN_PASSWORD_LENGTH} from '../defaults/constants';
+import {JWT_OFFSET_SECONDS, LENGTH_50, LENGTH_8} from '../defaults/constants';
 import {
   ERR_MSG_CONFIRMATION_PASSWORD_REQUIRED,
   ERR_MSG_DIFFERENT_PASSWORDS,
@@ -21,6 +17,10 @@ import {
   ERR_MSG_PASSWORD_LENGTH,
   ERR_MSG_PASSWORD_REQUIRED
 } from '../defaults/field-validation-messages';
+import {AuthLoginDto} from '../model/auth-login.dto';
+import {AuthRegisterDto} from '../model/auth-register.dto';
+import {SpinnerService} from '../service/spinner.service';
+import {AuthTabIndex} from '../enums/auth-tab-index';
 
 @Component({
   selector: 'app-auth',
@@ -34,47 +34,45 @@ export class AuthComponent implements OnInit {
   registerForm: FormGroup;
   hidePassword: boolean = true;
   returnUrl: string;
+  syncError: boolean = false;
+  selectedTabIndex: AuthTabIndex = AuthTabIndex.LOGIN;
 
   constructor(activatedRoute: ActivatedRoute,
-              private formBuilder: FormBuilder,
-              private router: Router,
-              private toast: ToastrService,
+              private sessionService: SessionService,
               private authService: AuthService,
               private spinnerService: SpinnerService,
-              private sessionService: SessionService) {
-
-    const accessToken: string = this.sessionService.getAccessToken();
-    const refreshToken: string = this.sessionService.getRefreshToken();
-
-    if (accessToken && refreshToken && !this.jwtHelper.isTokenExpired(refreshToken, JWT_OFFSET_SECONDS)) {
-      this.spinnerService.show();
-      this.sessionService.sync()
-        .subscribe(() => {
-          console.log('Sync completed');
-          this.router.navigate(['/home']);
-          this.spinnerService.close();
-        }, () => this.spinnerService.close());
-    } else {
-      localStorage.clear();
-    }
+              private toastrService: ToastrService,
+              private formBuilder: FormBuilder,
+              private router: Router) {
 
     this.returnUrl = activatedRoute.snapshot.queryParams['returnUrl'];
+    this.syncError = activatedRoute.snapshot.queryParams['syncError'];
 
   }
 
   ngOnInit(): void {
-    this.loginForm = this.formBuilder.group({
-      email: [undefined, [Validators.required, Validators.email]],
-      password: [undefined, [Validators.required]],
-    });
+    const accessToken = this.sessionService.getAccessToken();
+    const refreshToken = this.sessionService.getRefreshToken();
 
-    this.registerForm = this.formBuilder.group({
-      email: [undefined, [Validators.required, Validators.email, Validators.maxLength(MAX_EMAIL_LENGTH)]],
-      password: [undefined, [Validators.required, Validators.minLength(MIN_PASSWORD_LENGTH), Validators.maxLength(MAX_PASSWORD_LENGTH)]],
-      confirmPassword: [undefined, [Validators.required, Validators.minLength(MIN_PASSWORD_LENGTH), Validators.maxLength(MAX_PASSWORD_LENGTH)]]
-    }, {
-      validators: [PasswordValidator.validate]
-    });
+    if (!this.syncError && accessToken && refreshToken && !this.jwtHelper.isTokenExpired(refreshToken, JWT_OFFSET_SECONDS)) {
+      console.log('Valid tokens, redirect /home');
+      this.router.navigate(['/home']);
+    } else {
+      localStorage.clear();
+
+      this.loginForm = this.formBuilder.group({
+        email: [undefined, [Validators.required, Validators.email]],
+        password: [undefined, [Validators.required]],
+      });
+
+      this.registerForm = this.formBuilder.group({
+        email: [undefined, [Validators.required, Validators.email, Validators.maxLength(LENGTH_50)]],
+        password: [undefined, [Validators.required, Validators.minLength(LENGTH_8), Validators.maxLength(LENGTH_50)]],
+        confirmPassword: [undefined, [Validators.required, Validators.minLength(LENGTH_8), Validators.maxLength(LENGTH_50)]]
+      }, {
+        validators: [PasswordValidator.validate]
+      });
+    }
   }
 
   getLoginEmailErrorMessage(): string {
@@ -127,15 +125,13 @@ export class AuthComponent implements OnInit {
     if (this.loginForm.valid) {
       this.spinnerService.show();
 
-      const loginRequest: LoginRequest = new LoginRequest();
-      loginRequest.email = this.loginForm.value.email;
-      loginRequest.password = this.loginForm.value.password;
+      const authLogin: AuthLoginDto = {
+        email: this.loginForm.value.email,
+        password: this.loginForm.value.password
+      }
 
-      return this.authService.login(loginRequest)
-        .pipe(concatMap(() => this.sessionService.sync()))
+      return this.authService.login(authLogin)
         .subscribe(() => {
-          console.log('Sync completed');
-
           const reuseStrategy: CustomReuseStrategy = this.router.routeReuseStrategy as CustomReuseStrategy;
           reuseStrategy.routesToCache = ['home'];
           reuseStrategy.storedRouteHandles = new Map<string, DetachedRouteHandle>();
@@ -150,15 +146,19 @@ export class AuthComponent implements OnInit {
     if (this.registerForm.valid) {
       this.spinnerService.show();
 
-      const registerRequest: RegisterRequest = new RegisterRequest();
-      registerRequest.email = this.registerForm.value.email;
-      registerRequest.password = this.registerForm.value.password;
-      registerRequest.confirmPassword = this.registerForm.value.confirmPassword;
+      const authRegister: AuthRegisterDto = {
+        email: this.registerForm.value.email,
+        password: this.registerForm.value.password,
+        confirmPassword: this.registerForm.value.confirmPassword
+      }
 
-      return this.authService.register(registerRequest)
+      return this.authService.register(authRegister)
         .subscribe(user => {
-          this.toast.success('Registration successful');
+          this.toastrService.success('Registration successful');
           this.spinnerService.close();
+          this.selectedTabIndex = AuthTabIndex.LOGIN;
+          this.loginForm.get('email').setValue(authRegister.email);
+          this.loginForm.get('password').setValue(authRegister.password);
         }, () => this.spinnerService.close());
     }
   }
@@ -169,14 +169,20 @@ class PasswordValidator {
   static validate(control: AbstractControl): ValidationErrors | null {
     const password: string = control.get('password')?.value;
     const confirmPassword: string = control.get('confirmPassword')?.value;
+
     if (!password || !confirmPassword) {
+      control.get('confirmPassword')?.setErrors(null);
       return null;
     }
+
     if (password !== confirmPassword) {
       control.get('confirmPassword')?.setErrors({not_the_same: true});
       return ({not_the_same: true});
     }
+
+    control.get('confirmPassword')?.setErrors(null);
     return null;
   }
+
 }
